extract rdfa-parsing functionality out of say-parser and into seperate function
