+++ addon/components/_private/rdfa-editor/property-editor/modal.ts
@@ -1,5 +1,5 @@
 import Component from '@glimmer/component';
-import { AttributeProperty } from '@lblod/ember-rdfa-editor/core/say-parser';
+import { AttributeProperty } from '@lblod/ember-rdfa-editor/core/rdfa-annotator';
 import { localCopy } from 'tracked-toolbox';
 
 type Args = {
