+++ addon/components/_private/rdfa-editor/relationship-editor/index.ts
@@ -21,7 +21,7 @@ import {
   ExternalProperty,
   ExternalPropertyObject,
   Property,
-} from '@lblod/ember-rdfa-editor/core/say-parser';
+} from '@lblod/ember-rdfa-editor/core/rdfa-annotator';
 
 type Args = {
   controller?: SayController;
