+++ addon/utils/_private/html-utils.ts
@@ -1,4 +1,5 @@
 import { PNode, ProseParser } from '@lblod/ember-rdfa-editor';
+import { annotate } from '@lblod/ember-rdfa-editor/core/rdfa-annotator';
 import { Schema } from 'prosemirror-model';
 import HTMLInputParser from './html-input-parser';
 
@@ -13,8 +14,8 @@ export function htmlToDoc(
   const parsed = domParser.parseFromString(cleanedHTML, 'text/html').body;
   const documentDiv = parsed.querySelector('div[data-say-document="true"]');
   let doc: PNode;
-  console.log('gonna parse with', parser);
   if (documentDiv) {
+    annotate(documentDiv);
     doc = parser.parse(documentDiv, {
       preserveWhitespace: true,
       topNode: schema.nodes.doc.create({
@@ -29,6 +30,7 @@ export function htmlToDoc(
       }),
     });
   } else {
+    annotate(parsed);
     doc = parser.parse(parsed, {
       preserveWhitespace: true,
     });
